+++ CHANGELOG.md
@@ -16,3 +16,4 @@
 - [fix: fix lb configuration on bootstrap step.](https://github.com/Tencent/spring-cloud-tencent/issues/1673)
 - [feat:support 2024.0.2.](https://github.com/Tencent/spring-cloud-tencent/issues/1674)
 - [feat:support fault injection.](https://github.com/Tencent/spring-cloud-tencent/pull/1672)
+- [feat: support config event and monitor address list.](https://github.com/Tencent/spring-cloud-tencent/pull/1675)